refactored gui

+++ data/mnist_data_handler.py
@@ -18,8 +18,8 @@ def get_basic_data(categorical: bool = False) -> Tuple[Tuple[Any, Any], Tuple[An
     x_test = x_test.reshape(x_test.shape[0], img_size, 1)
     input_shape = (img_size, 1)
 
-    x_train = x_train.astype('float32')
-    x_test = x_test.astype('float32')
+    x_train = x_train.astype("float32")
+    x_test = x_test.astype("float32")
     x_train /= 255
     x_test /= 255
 
@@ -168,9 +168,9 @@ def split_mnist_data(class_selection: List[int] = None):
         np.savez("%s/mnist_train_split" % data_path, processed_separated_train_data)
         np.savez("%s/mnist_test_split" % data_path, processed_separated_test_data)
     else:
-        np.savez("%s/mnist_train_split_%s" % (data_path, ''.join(str(e) + '_' for e in class_selection)),
+        np.savez("%s/mnist_train_split_%s" % (data_path, "".join(str(e) + "_" for e in class_selection)),
                  processed_separated_train_data)
-        np.savez("%s/mnist_test_split_%s" % (data_path, ''.join(str(e) + '_' for e in class_selection)),
+        np.savez("%s/mnist_test_split_%s" % (data_path, "".join(str(e) + "_" for e in class_selection)),
                  processed_separated_test_data)
 
     logging.info("saved split data to \"%s\"" % data_path)
